foundation for webpack migration
