+++ package.json
@@ -6,13 +6,13 @@
   "scripts": {
     "clean": "rimraf public",
     "build:dll": "rimraf public && webpack --config webpack.dll.js",
-    "build": "webpack --config webpack.dev.js",
-    "build:watch": "webpack --config webpack.dev.js --watch --progress",
+    "build": "webpack --mode development --config webpack.config.js --progress",
+    "build:watch": "webpack --mode development --config webpack.config.js --watch --progress",
     "run": "node --experimental-modules server.mjs",
     "run:dev": "nodemon --experimental-modules server.mjs --watch utils/ --watch events/ --watch server.js",
     "dev": "npm-run-all --parallel build:watch run:dev",
     "eslint": "eslint ./src",
-    "build:prod": "rimraf public && webpack --env prod --config webpack.dev.js",
+    "build:prod": "rimraf public && webpack --mode production --config webpack.config.js",
     "test": "node_modules/jest/bin/jest.js",
     "test:fe": "node_modules/jest/bin/jest.js src/*",
     "test:silent": "NODE_HIDECONSOLE=true npm run test"
@@ -74,7 +74,7 @@
   },
   "devDependencies": {
     "@babel/core": "7.0.0-beta.47",
-    "@babel/plugin-proposal-class-properties": "7.0.0-beta.47",
+    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.47",
     "@babel/plugin-transform-modules-commonjs": "7.0.0-beta.52",
     "@babel/preset-env": "7.0.0-beta.47",
     "@babel/preset-react": "7.0.0-beta.47",
@@ -103,7 +103,7 @@
     "express": "4.14.1",
     "express-static-gzip": "0.3.2",
     "faker": "4.1.0",
-    "file-loader": "0.10.1",
+    "file-loader": "1.1.11",
     "fs": "0.0.1-security",
     "html-webpack-plugin": "2.30.1",
     "identity-obj-proxy": "3.0.0",
@@ -127,7 +127,8 @@
     "uglifyjs-webpack-plugin": "1.2.2",
     "units-css": "0.4.0",
     "url-loader": "1.1.0",
-    "webpack": "3.10.0",
-    "webpack-bundle-analyzer": "2.9.1"
+    "webpack": "4.16.5",
+    "webpack-bundle-analyzer": "2.13.1",
+    "webpack-cli": "^3.1.0"
   }
 }
